requirements.txt
requests
plyer
PyGetWindow
pydub
pywin32; sys_platform == "win32"
pyobjc; sys_platform == "darwin"



# Core dependencies  
numpy>=2.0,<2.1  
cython>=3.0.0  
scipy>=1.13.0  
torch>=2.3  
torchaudio>=2.3.0  
soundfile>=0.12.0  
librosa>=0.11.0  
inflect>=5.6.0  
tqdm>=4.64.1  
anyascii>=0.3.0  
pyyaml>=6.0  
fsspec[http]>=2023.6.0  
packaging>=23.1  
typing_extensions>=4.10  
  
# Inference  
pysbd>=0.3.4  
  
# Training  
matplotlib>=3.8.4  
  
# Coqui stack  
coqui-tts-trainer>=0.2.0,<0.3.0  
coqpit-config>=0.2.0,<0.3.0  
monotonic-alignment-search>=0.1.0  
  
# Gruut + supported languages  
gruut[de,es,fr]>=2.4.0  
  
# Tortoise  
einops>=0.6.0  
transformers>=4.47.0  
  
# Bark  
encodec>=0.1.1  
  
# XTTS  
num2words>=0.5.14  
spacy[ja]>=3.8,<4